pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "oauthlib>=3.3.1",
    "anthropic>=0.55.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.174.0",
    "google-auth-oauthlib>=1.2.2",
    "pandas>=2.3.0",
    "pytz>=2025.2",
    "flask-cors>=6.0.1",
    "google-auth>=2.40.3",
]
